Move StatBox out of unnecessary stats folder
